+++ packages/dayjs-helpers/src/utils/calc-last-date.ts
@@ -1,0 +1,11 @@
+import dayjs from "dayjs";
+
+import type { NullishDayjsLike } from "../types";
+import { _sortDates } from "./sort-dates";
+
+export const _calcLastDate =
+  (_dayjs: typeof dayjs) => (values: NullishDayjsLike[]) => {
+    const sortedValues = _sortDates(_dayjs)(values);
+
+    return sortedValues[sortedValues.length - 1];
+  };