+++ packages/dayjs-helpers/src/zod/datetime.ts
@@ -1,0 +1,7 @@
+import { RequiredStringSchema } from "@keanugrieves/zod-helpers";
+import dayjs from "dayjs";
+
+export const DatetimeSchema = RequiredStringSchema.refine(
+  (v) => dayjs(v).isValid(),
+  "Must be a valid datetime"
+);