+++ packages/zod-helpers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @keanugrieves/zod-helpers
 
+## 2.1.3
+
+### Patch Changes
+
+- Remove tests
+
+## 2.1.2
+
+### Patch Changes
+
+- Convert to module
+
 ## 2.1.1
 
 ### Patch Changes